experiments/dqn/cartpole.yaml seems working now

+++ hw3/cs285/infrastructure/replay_buffer.py
@@ -46,11 +46,11 @@ class ReplayBuffer:
             )
         """
         if isinstance(reward, (float, int)):
-            reward = np.array(reward)
+            reward = np.array(reward, dtype=np.float32)
         if isinstance(done, bool):
-            done = np.array(done)
+            done = np.array(done, dtype=np.uint8)
         if isinstance(action, int):
-            action = np.array(action, dtype=np.int64)
+            action = np.array(action, dtype=np.uint8)  # for discrete actions, 256 should be enough
 
         if self.observations is None:
             self.observations = np.empty(
